revert speed controls to int

+++ shaders/effects/classicNoisedeck/tunnel/definition.js
@@ -63,7 +63,7 @@ export default new Effect({
       }
     },
     speed: {
-      type: "float",
+      type: "int",
       default: 1,
       uniform: "speed",
       min: -5,
